+++ php-dev-helper/includers/include_common.sh
@@ -1,19 +1,18 @@
 #!/usr/bin/env sh
-# shellcheck disable=SC1090
+
+# Disable source following.
+#   shellcheck disable=SC1090,SC1091
+
 . "${COMMON_MODULES_DIR}/options.sh"
-# shellcheck disable=SC1090
 . "${COMMON_MODULES_DIR}/functions.sh"
-# shellcheck disable=SC1090
 . "${COMMON_MODULES_DIR}/commands.sh"
-# shellcheck disable=SC1090
 . "${COMMON_MODULES_DIR}/settings.sh"
-# shellcheck disable=SC1090
 . "${COMMON_MODULES_DIR}/tmp.sh"
+
 # Load PTS_AUX_DEV_MODULE module if present
 if [ -e "${COMMON_MODULES_DIR}/${PTS_AUX_DEV_MODULE}" ]; then
     console_debug "Processing module '${PTS_AUX_DEV_MODULE}'"
-    # shellcheck disable=SC1090
     . "${COMMON_MODULES_DIR}/${PTS_AUX_DEV_MODULE}" && console_debug "Module '${PTS_AUX_DEV_MODULE}' processed"
 else
-    console_debug "Aux dev module '${PTS_AUX_DEV_MODULE}' not found"
+    console_debug "Module '${PTS_AUX_DEV_MODULE}' not found"
 fi