+++ php-dev-helper/includers/include_pts.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env sh
 
-# shellcheck disable=SC1090
+# Disable source following.
+#   shellcheck disable=SC1090,SC1091
+
 . "${APPS_MODULES_DIR}/pts.sh"               # Image builder module
 
 ### Read options