+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 
 ### Added
+- `--source-dir` option
 - tests framework
 
 ### Fixed
@@ -10,11 +11,10 @@
 ### Modified
 - shunit2
 
-### Update
-- README.md
-
 ### Updated
+- readme file
 - Changelog
+- README.md
 
 
 <a name="0.5.3"></a>